requirements.txt
python-telegram-bot==13.9
requests>=2.20.0